NyayaMitra/backend/requirements.txt
# ── Core ──────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-dotenv==1.0.1

# ── AI / LLM ─────────────────────────────────────────
chromadb==0.5.23
openai==1.58.1
groq==0.13.1

# ── HTTP / Files ─────────────────────────────────────
httpx==0.28.1
aiofiles==24.1.0
requests==2.32.3

# ── Scraping / PDF ───────────────────────────────────
beautifulsoup4==4.12.3
pdfplumber==0.11.4

# ── ML ───────────────────────────────────────────────
scikit-learn==1.6.1
pandas==2.2.3
numpy==2.2.1

# ── Integrations ─────────────────────────────────────
python-telegram-bot==21.9
supabase==2.11.0
web3==7.6.0
